+++ TopDownCalculator/src/net/codejava/CalculatorPortBindingSkeleton.java
@@ -31,11 +31,11 @@ public class CalculatorPortBindingSkeleton implements net.codejava.Calculator, o
         org.apache.axis.description.FaultDesc _fault;
         org.apache.axis.description.ParameterDesc [] _params;
         _params = new org.apache.axis.description.ParameterDesc [] {
-            new org.apache.axis.description.ParameterDesc(new javax.xml.namespace.QName("", "arg0"), org.apache.axis.description.ParameterDesc.IN, new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "int"), int.class, false, false), 
-            new org.apache.axis.description.ParameterDesc(new javax.xml.namespace.QName("", "arg1"), org.apache.axis.description.ParameterDesc.IN, new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "int"), int.class, false, false), 
+            new org.apache.axis.description.ParameterDesc(new javax.xml.namespace.QName("", "arg0"), org.apache.axis.description.ParameterDesc.IN, new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "double"), double.class, false, false), 
+            new org.apache.axis.description.ParameterDesc(new javax.xml.namespace.QName("", "arg1"), org.apache.axis.description.ParameterDesc.IN, new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "double"), double.class, false, false), 
         };
         _oper = new org.apache.axis.description.OperationDesc("add", _params, new javax.xml.namespace.QName("", "return"));
-        _oper.setReturnType(new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "int"));
+        _oper.setReturnType(new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "double"));
         _oper.setElementQName(new javax.xml.namespace.QName("http://codejava.net/", "add"));
         _oper.setSoapAction("");
         _myOperationsList.add(_oper);
@@ -44,11 +44,11 @@ public class CalculatorPortBindingSkeleton implements net.codejava.Calculator, o
         }
         ((java.util.List)_myOperations.get("add")).add(_oper);
         _params = new org.apache.axis.description.ParameterDesc [] {
-            new org.apache.axis.description.ParameterDesc(new javax.xml.namespace.QName("", "arg0"), org.apache.axis.description.ParameterDesc.IN, new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "int"), int.class, false, false), 
-            new org.apache.axis.description.ParameterDesc(new javax.xml.namespace.QName("", "arg1"), org.apache.axis.description.ParameterDesc.IN, new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "int"), int.class, false, false), 
+            new org.apache.axis.description.ParameterDesc(new javax.xml.namespace.QName("", "arg0"), org.apache.axis.description.ParameterDesc.IN, new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "double"), double.class, false, false), 
+            new org.apache.axis.description.ParameterDesc(new javax.xml.namespace.QName("", "arg1"), org.apache.axis.description.ParameterDesc.IN, new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "double"), double.class, false, false), 
         };
         _oper = new org.apache.axis.description.OperationDesc("sub", _params, new javax.xml.namespace.QName("", "return"));
-        _oper.setReturnType(new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "int"));
+        _oper.setReturnType(new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "double"));
         _oper.setElementQName(new javax.xml.namespace.QName("http://codejava.net/", "sub"));
         _oper.setSoapAction("");
         _myOperationsList.add(_oper);
@@ -57,11 +57,11 @@ public class CalculatorPortBindingSkeleton implements net.codejava.Calculator, o
         }
         ((java.util.List)_myOperations.get("sub")).add(_oper);
         _params = new org.apache.axis.description.ParameterDesc [] {
-            new org.apache.axis.description.ParameterDesc(new javax.xml.namespace.QName("", "arg0"), org.apache.axis.description.ParameterDesc.IN, new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "int"), int.class, false, false), 
-            new org.apache.axis.description.ParameterDesc(new javax.xml.namespace.QName("", "arg1"), org.apache.axis.description.ParameterDesc.IN, new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "int"), int.class, false, false), 
+            new org.apache.axis.description.ParameterDesc(new javax.xml.namespace.QName("", "arg0"), org.apache.axis.description.ParameterDesc.IN, new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "double"), double.class, false, false), 
+            new org.apache.axis.description.ParameterDesc(new javax.xml.namespace.QName("", "arg1"), org.apache.axis.description.ParameterDesc.IN, new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "double"), double.class, false, false), 
         };
         _oper = new org.apache.axis.description.OperationDesc("div", _params, new javax.xml.namespace.QName("", "return"));
-        _oper.setReturnType(new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "int"));
+        _oper.setReturnType(new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "double"));
         _oper.setElementQName(new javax.xml.namespace.QName("http://codejava.net/", "div"));
         _oper.setSoapAction("");
         _myOperationsList.add(_oper);
@@ -70,11 +70,11 @@ public class CalculatorPortBindingSkeleton implements net.codejava.Calculator, o
         }
         ((java.util.List)_myOperations.get("div")).add(_oper);
         _params = new org.apache.axis.description.ParameterDesc [] {
-            new org.apache.axis.description.ParameterDesc(new javax.xml.namespace.QName("", "arg0"), org.apache.axis.description.ParameterDesc.IN, new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "int"), int.class, false, false), 
-            new org.apache.axis.description.ParameterDesc(new javax.xml.namespace.QName("", "arg1"), org.apache.axis.description.ParameterDesc.IN, new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "int"), int.class, false, false), 
+            new org.apache.axis.description.ParameterDesc(new javax.xml.namespace.QName("", "arg0"), org.apache.axis.description.ParameterDesc.IN, new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "double"), double.class, false, false), 
+            new org.apache.axis.description.ParameterDesc(new javax.xml.namespace.QName("", "arg1"), org.apache.axis.description.ParameterDesc.IN, new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "double"), double.class, false, false), 
         };
         _oper = new org.apache.axis.description.OperationDesc("mul", _params, new javax.xml.namespace.QName("", "return"));
-        _oper.setReturnType(new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "int"));
+        _oper.setReturnType(new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "double"));
         _oper.setElementQName(new javax.xml.namespace.QName("http://codejava.net/", "mul"));
         _oper.setSoapAction("");
         _myOperationsList.add(_oper);
@@ -91,27 +91,27 @@ public class CalculatorPortBindingSkeleton implements net.codejava.Calculator, o
     public CalculatorPortBindingSkeleton(net.codejava.Calculator impl) {
         this.impl = impl;
     }
-    public int add(int arg0, int arg1) throws java.rmi.RemoteException
+    public double add(double arg0, double arg1) throws java.rmi.RemoteException
     {
-        int ret = impl.add(arg0, arg1);
+        double ret = impl.add(arg0, arg1);
         return ret;
     }
 
-    public int sub(int arg0, int arg1) throws java.rmi.RemoteException
+    public double sub(double arg0, double arg1) throws java.rmi.RemoteException
     {
-        int ret = impl.sub(arg0, arg1);
+        double ret = impl.sub(arg0, arg1);
         return ret;
     }
 
-    public int div(int arg0, int arg1) throws java.rmi.RemoteException
+    public double div(double arg0, double arg1) throws java.rmi.RemoteException
     {
-        int ret = impl.div(arg0, arg1);
+        double ret = impl.div(arg0, arg1);
         return ret;
     }
 
-    public int mul(int arg0, int arg1) throws java.rmi.RemoteException
+    public double mul(double arg0, double arg1) throws java.rmi.RemoteException
     {
-        int ret = impl.mul(arg0, arg1);
+        double ret = impl.mul(arg0, arg1);
         return ret;
     }
